+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/doeverything/main/controllers/DoEverythingController.scala
@@ -205,7 +205,7 @@ class DoEverythingController @Inject()(
     "bar"
   }
 
-  post("/formPost") { formPost: FormPostRequest =>
+  post("/formPost/:card_id") { formPost: FormPostRequest =>
     formPost.name
   }